options: remove CONF_TYPE_POSITION
This was the option parser for the off_t C type. These days, off_t is
always int64_t, so replace all its uses by int64_t and CONF_TYPE_INT64.

Fix the --sstep option. It used CONF_TYPE_INT with an off_t variable,
which will result in invalid memory accesses. Make it use type double
instead, which seems to make more sense for this option.

+++ mplayer.c
@@ -204,8 +204,8 @@ static const char av_desync_help_text[] = _(
 static int drop_frame_cnt; // total number of dropped frames
 
 // seek:
-static off_t seek_to_byte;
-static off_t step_sec;
+static int64_t seek_to_byte;
+static double step_sec;
 
 static m_time_size_t end_at = { .type = END_AT_NONE, .pos = 0 };
